+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [Unreleased]
 
+- Improved graphics sizes
+- Changed `E[q]`, `E[ω]` to sample size tooltip
+- Removed model type `Undeveloped`
+
 
 ## [v0.1.0] First unstable (2025-05-09)